I've made more changes to SideKickParsedData.java for making the highlighted node in the tree follow the caret in the text editing pane.  It works like this:
1. If the caret is within the bounds of the deepest child/leaf node, that
node is highlighted.

2. If the caret is between child nodes, the next child node is highlighted.

3. As an exception to #2, if the caret is between the start of a node and
the first child node, the node itself is highlighted.

4. If the caret is in a node, but after all child nodes, the node itself
is highlighted.

The reason for #3 is that without the exception, if you use your mouse and
click on the start of a node (the name of a method in java, for example),
the highlight in the tree would move to the first child.  There really isn't
enough granularity in the positions provided by Asset to do much more.