Speed up file importing a lot.
Several things in this change:
. cache some things in VPTFile so that it doesn't need to create new strings as much.
. rework the addTree() method (now called importFiles()) so that it doesn't rely so much on constructPath(). constructPath() is very expensive, so it's better to avoid it whenever possible.
. change some of the lists in the Importer implementation to use hash maps, to make lookup times lower.
. only fire tree notifications for nodes that are already in the tree; the old code would notify the tree about every single node being added, which took forever.

In the process, some code was reorganized to be more generic, make the implementation of a few methods simpler. the AutoReimporter code was also changed to use the new methods (which are based on its old "importFiles()" implementation).

There are still possible improvements (e.g., do a binary search in Importer.findChild() instead of a linear search), but with this implementation importing a directory with 41k files (of which 21k were imported) took about 5s on my machine, against between 5 and 6 minutes earlier. According to JProfiler, most of the time in the current implementation is spent inside the VFS layer (meaning in jEdit code, not in PV), sorting the data returned when listing directories.