SF FR #1515218: save different filter settings for "add files" and for the project itself.
Turns out I had already written most of the needed code as part of another feature. This just uses that code to keep two different filter settings: one for the "add files" dialog (which falls back to the project filter if not set), and a second for the project (used during re-import, for example).