Added line-based results using the Highlighter package. Added search options in the dockable: - Whether to show line-based results - The maximal number of results to show For the highlighter to work, a highlight-scorer must be created based on the query, so the result processor also receives the query with each result.
LineIndexImpl can almost be considered deprecated, since an up-to-date line-based results can be retrieved using IndexImpl with the Highlighter. But, using the Highlighter for this requires a lot of memory, since the entire file must be in memory and an array of line start positions must be maintained for this, so LineIndexImpl can trade memory for performance and index size on disk.